requirements.txt
openpyxl>=3.1.2
rapidfuzz>=3.9.0
PyYAML>=6.0.1
pytest>=8.0.0
